+++ include/rtthread.h
@@ -124,23 +124,6 @@ void rt_timer_enter_sethook(void (*hook)(struct rt_timer *timer));
 void rt_timer_exit_sethook(void (*hook)(struct rt_timer *timer));
 #endif
 
-#ifdef RT_USING_CPUTIME
-void rt_cputime_timer_init(void);
-rt_err_t rt_cputime_timer_detach(rt_cputime_timer_t timer);
-#ifdef RT_USING_HEAP
-
-void rt_cputime_timer_create(rt_cputime_timer_t timer,
-                             const char *name,
-                             void (*timeout)(void *parameter),
-                             void *parameter,
-                             rt_uint64_t time,
-                             rt_uint8_t flag);
-rt_err_t rt_cputime_timer_delete(rt_cputime_timer_t timer);
-#endif
-rt_err_t rt_cputime_timer_start(rt_cputime_timer_t timer);
-rt_err_t rt_cputime_timer_stop(rt_cputime_timer_t timer);
-rt_err_t rt_cputime_timer_control(rt_cputime_timer_t timer, int cmd, void *arg);
-#endif /* RT_USING_CPUTIME */
 /**@}*/
 
 /**
@@ -180,12 +163,6 @@ rt_err_t rt_thread_control(rt_thread_t thread, int cmd, void *arg);
 rt_err_t rt_thread_suspend(rt_thread_t thread);
 rt_err_t rt_thread_suspend_with_flag(rt_thread_t thread, int suspend_flag);
 rt_err_t rt_thread_resume(rt_thread_t thread);
-#ifdef RT_USING_CPUTIME
-rt_err_t rt_cputime_sleep(rt_uint64_t tick);
-rt_err_t rt_cputime_ndelay(rt_uint64_t ns);
-rt_err_t rt_cputime_udelay(rt_uint64_t us);
-rt_err_t rt_cputime_mdelay(rt_uint64_t ms);
-#endif /* RT_USING_CPUTIME */
 #ifdef RT_USING_SMART
 rt_err_t rt_thread_wakeup(rt_thread_t thread);
 void rt_thread_wakeup_set(struct rt_thread *thread, rt_wakeup_func_t func, void* user_data);
